+++ src/Services/ReviewServices/reviewapi.js
@@ -3,6 +3,11 @@ import API from "../../config/axios";
 // ✅ Post a Review
 export const postReview = async (eventId, {rating, comment }) => {
   try {
+    const token = localStorage.getItem("accessToken");
+    if(!token){
+      console.log("User not logged in");
+      return
+    }
     const res = await API.post(`/api/v1/reviews/post-review/${eventId}`, {
             rating,
             comment
@@ -24,8 +29,13 @@ export const getEventReviews = async (eventId) => {
 };
 
 // ✅ Edit a Review
-export const editReview = async (reviewId, {rating, comment}) => {
+export const updateReview = async (reviewId, {rating, comment}) => {
   try {
+    const token = localStorage.getItem("accessToken");
+    if(!token){
+      console.log("User not logged in");
+      return
+    }
     const res = await API.patch(`/api/v1/reviews/edit/${reviewId}`, {
       rating,
       comment
@@ -39,6 +49,11 @@ export const editReview = async (reviewId, {rating, comment}) => {
 // ✅ Delete a Review
 export const deleteReview = async (reviewId) => {
   try {
+    const token = localStorage.getItem("accessToken");
+    if(!token){
+      console.log("User not logged in");
+      return
+    }
     const res = await API.delete(`/api/v1/reviews/delete-review/${reviewId}`);
     return res.data;
   } catch (error) {
